fix some warnings by casting safely. (trivial)
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@4064 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ util/romcc/romcc.c
@@ -20340,8 +20340,8 @@ static void scc_add_sedge(struct compile_state *state, struct scc_state *scc,
 	struct ssa_edge *sedge)
 {
 	if (state->compiler->debug & DEBUG_SCC_TRANSFORM2) {
-		fprintf(state->errout, "adding sedge: %5d (%4d -> %5d)\n",
-			sedge - scc->ssa_edges,
+		fprintf(state->errout, "adding sedge: %5ld (%4d -> %5d)\n",
+			(long)(sedge - scc->ssa_edges),
 			sedge->src->def->id,
 			sedge->dst->def->id);
 	}
@@ -20350,8 +20350,8 @@ static void scc_add_sedge(struct compile_state *state, struct scc_state *scc,
 		(sedge->work_prev != sedge)) {
 
 		if (state->compiler->debug & DEBUG_SCC_TRANSFORM2) {
-			fprintf(state->errout, "dupped sedge: %5d\n",
-				sedge - scc->ssa_edges);
+			fprintf(state->errout, "dupped sedge: %5ld\n",
+				(long)(sedge - scc->ssa_edges));
 		}
 		return;
 	}
@@ -23843,12 +23843,12 @@ static long get_const_pool_ref(
 	long ref;
 	ref = next_label(state);
 	fprintf(fp, ".section \"" DATA_SECTION "\"\n");
-	fprintf(fp, ".balign %d\n", align_of_in_bytes(state, ins->type));
+	fprintf(fp, ".balign %ld\n", align_of_in_bytes(state, ins->type));
 	fprintf(fp, "L%s%lu:\n", state->compiler->label_prefix, ref);
 	print_const(state, ins, fp);
 	fill_bytes = bits_to_bytes(size - size_of(state, ins->type));
 	if (fill_bytes) {
-		fprintf(fp, ".fill %d, 1, 0\n", fill_bytes);
+		fprintf(fp, ".fill %ld, 1, 0\n", fill_bytes);
 	}
 	fprintf(fp, ".section \"" TEXT_SECTION "\"\n");
 	return ref;
@@ -24657,7 +24657,7 @@ static void print_sdecl(struct compile_state *state,
 	struct triple *ins, FILE *fp)
 {
 	fprintf(fp, ".section \"" DATA_SECTION "\"\n");
-	fprintf(fp, ".balign %d\n", align_of_in_bytes(state, ins->type));
+	fprintf(fp, ".balign %ld\n", align_of_in_bytes(state, ins->type));
 	fprintf(fp, "L%s%lu:\n", 
 		state->compiler->label_prefix, (unsigned long)(ins->u.cval));
 	print_const(state, MISC(ins, 0), fp);
